added product data to api
