+++ const.go
@@ -3,5 +3,5 @@ package gosupreme
 // Endpoint
 const (
 	StockURL = "https://supremenewyork.com/mobile_stock.json"
-	StyleURL = "https://supremenewyork.com/shop/%d.json"
+	DataURL  = "https://supremenewyork.com/shop/%d.json"
 )
