+++ style.go
@@ -1,23 +1,11 @@
 package gosupreme
 
-type Data struct {
-	Styles                  Styles      `json:"styles"`
-	Description             string      `json:"description"`
-	CanAddStyles            bool        `json:"can_add_styles"`
-	CanBuyMultiple          bool        `json:"can_buy_multiple"`
-	Ino                     string      `json:"ino"`
-	CodBlocked              bool        `json:"cod_blocked"`
-	CanadaBlocked           bool        `json:"canada_blocked"`
-	MexicoBlocked           bool        `json:"mexico_blocked"`
-	PurchasableQty          int         `json:"purchasable_qty"`
-	NewItem                 bool        `json:"new_item"`
-	Apparel                 bool        `json:"apparel"`
-	Handling                int         `json:"handling"`
-	NoFreeShipping          bool        `json:"no_free_shipping"`
-	CanBuyMultipleWithLimit int         `json:"can_buy_multiple_with_limit"`
-	Tag                     interface{} `json:"tag"`
-}
+import (
+	"fmt"
+	"net/http"
+)
 
+// Represents the a product style
 type Style struct {
 	ID                    int         `json:"id"`
 	Name                  string      `json:"name"`
@@ -38,6 +26,7 @@ type Style struct {
 
 type Styles []*Style
 
+// Represents a product style size
 type Size struct {
 	Name       string `json:"name"`
 	ID         int    `json:"id"`
@@ -45,3 +34,26 @@ type Size struct {
 }
 
 type Sizes []*Size
+
+func (d *Data) String() string {
+	return fmt.Sprintf("Description: %s\n", d.Description)
+}
+
+func (s *Supreme) FetchProductData(p *Product) (*Data, error) {
+	s.l.Printf("Fetching product data for %d (%s)\n", p.ID, p.Name)
+
+	// create a new http GET request
+	req, err := http.NewRequest("GET", fmt.Sprintf(DataURL, p.ID), nil)
+	if err != nil {
+		return nil, err
+	}
+
+	// make the http request and save its response
+	var data *Data
+	err = s.makeRequest(req, &data)
+	if err != nil {
+		return nil, err
+	}
+
+	return data, nil
+}
